Simple Map Generator Module
First pass at a simple map generator using Axial Coordinates.

Can be run as a standalone program to test different configurations of values. More intended to be imported into another program and run by passing values.

Could be possible to have this populate an instance of a class rather than populating an array of coordinates.

For some reason the map will include extra coordinates on the x axis. Haven't taken the time to figure that out yet.